# core-banking-ledger-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-ledger
  description: Banking Ledger Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/ledger
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/transactions/{transactionId}/line-sepa-transfer:
    get:
      tags:
        - Transaction Line SEPA Transfer
      summary: Get SEPA Transfer Line
      description: Retrieve the SEPA transfer line record associated with the specified transaction.
      operationId: getSepaTransferLine
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the SEPA transfer line record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionLineSepaTransferDTO'
        '404':
          description: SEPA transfer line record not found for this transaction
    put:
      tags:
        - Transaction Line SEPA Transfer
      summary: Update SEPA Transfer Line
      description: Update an existing SEPA transfer line record associated with the specified transaction.
      operationId: updateSepaTransferLine
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionLineSepaTransferDTO'
        required: true
      responses:
        '200':
          description: SEPA transfer line updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionLineSepaTransferDTO'
        '404':
          description: SEPA transfer line record not found for this transaction
    post:
      tags:
        - Transaction Line SEPA Transfer
      summary: Create SEPA Transfer Line
      description: Create a new SEPA transfer line record associated with the specified transaction.
      operationId: createSepaTransferLine
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionLineSepaTransferDTO'
        required: true
      responses:
        '201':
          description: SEPA transfer line record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionLineSepaTransferDTO'
        '400':
          description: Invalid SEPA transfer line data provided
    delete:
      tags:
        - Transaction Line SEPA Transfer
      summary: Delete SEPA Transfer Line
      description: Remove an existing SEPA transfer line record from the specified transaction.
      operationId: deleteSepaTransferLine
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: SEPA transfer line record deleted successfully
        '404':
          description: SEPA transfer line record not found for this transaction
...
  /api/v1/money:
    get:
      tags:
        - Money
      summary: List money
      description: Lists all money values
      operationId: listMoney
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: sort
          in: query
          description: Sort field
          required: false
          schema:
            type: string
        - name: direction
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Money list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseMoneyDTO'
    post:
      tags:
        - Money
      summary: Create money
      description: Creates a new money value
      operationId: createMoney
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoneyDTO'
        required: true
      responses:
        '200':
          description: Money created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoneyDTO'
        '400':
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MoneyDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MoneyDTO'
...
  /api/v1/money/{moneyId}:
    get:
      tags:
        - Money
      summary: Get money
      description: Gets a specific money value by ID
      operationId: getMoney
      parameters:
        - name: moneyId
          in: path
          description: ID of the money value
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Money retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoneyDTO'
        '404':
          description: Money not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MoneyDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MoneyDTO'
  /api/v1/money/currency/{currency}:
    get:
      tags:
        - Money
      summary: List money by currency
      description: Lists all money values with a specific currency
      operationId: listMoneyByCurrency
      parameters:
        - name: currency
          in: path
          description: Currency code
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: sort
          in: query
          description: Sort field
          required: false
          schema:
            type: string
        - name: direction
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Money list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseMoneyDTO'
...
components:
  schemas:
...
    TransactionLineSepaTransferDTO:
      required:
        - sepaDestinationIban
        - sepaEndToEndId
        - sepaOriginIban
        - sepaPaymentScheme
        - sepaTransactionStatus
        - transactionId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        transactionLineSepaId:
          type: string
          format: uuid
          readOnly: true
        transactionId:
          type: string
          format: uuid
        sepaEndToEndId:
          maxLength: 35
          minLength: 0
          type: string
        sepaRemittanceInfo:
          maxLength: 140
          minLength: 0
          type: string
        sepaOriginIban:
          type: string
        sepaOriginBic:
          type: string
        sepaDestinationIban:
          type: string
        sepaDestinationBic:
          type: string
        sepaTransactionStatus:
          type: string
          enum:
            - ACCP
            - RJCT
            - PDNG
        sepaCreditorId:
          maxLength: 35
          minLength: 0
          type: string
        sepaDebtorId:
          maxLength: 35
          minLength: 0
          type: string
        sepaInitiatingAgentBic:
          type: string
        sepaIntermediaryBic:
          type: string
        sepaTransactionPurpose:
          maxLength: 4
          minLength: 0
          type: string
        sepaRequestedExecutionDate:
          type: string
          format: date
        sepaExchangeRate:
          minimum: 0
          exclusiveMinimum: false
          type: number
        sepaFeeAmount:
          type: number
        sepaFeeCurrency:
          type: string
        sepaRecipientName:
          maxLength: 70
          minLength: 0
          type: string
        sepaRecipientAddress:
          maxLength: 140
          minLength: 0
          type: string
        sepaProcessingDate:
          type: string
          format: date-time
        sepaNotes:
          maxLength: 500
          minLength: 0
          type: string
        sepaPaymentScheme:
          type: string
          enum:
            - CORE
            - B2B
            - COR1
...
    MoneyDTO:
      required:
        - amount
        - currency
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        moneyId:
          type: string
          format: uuid
          readOnly: true
        amount:
          type: number
        currency:
          type: string
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
    PaginationResponseMoneyDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/MoneyDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...